requirements.txt
numpy
matplotlib
scikit-learn
pyyaml
tensorflow
